+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes qsf-integration project will be documented in this file.
 - add minTokenWeight for Spellcheck
 - implement a logic to track the reason for a changed query
 - implement a include for the json builder
+- jsonBuilder: implement an exists function
+- jsonBuilder: implement a function to create not existing paths pathsForceCreate
 
 ### Removed